refactor(data): moving and renaming mapper file
