+++ data/src/main/java/br/dev/singular/overview/data/repository/MediaRepository.kt
@@ -1,9 +1,9 @@
 package br.dev.singular.overview.data.repository
 
 import br.dev.singular.overview.data.local.source.IMediaRouteLocalDataSource
-import br.dev.singular.overview.data.model.toDomainModel
 import br.dev.singular.overview.data.network.source.DataResult
 import br.dev.singular.overview.data.network.source.IMediaRemoteDataSource
+import br.dev.singular.overview.data.util.mappers.toDomainModel
 import br.dev.singular.overview.domain.model.Media
 import br.dev.singular.overview.domain.model.MediaParam
 import br.dev.singular.overview.domain.repository.GetAll
