+++ data/src/test/java/br/dev/singular/overview/data/repository/SuggestionRepositoryTest.kt
@@ -1,8 +1,8 @@
 package br.dev.singular.overview.data.repository
 
 import br.dev.singular.overview.data.local.source.ISuggestionLocalDataSource
-import br.dev.singular.overview.data.model.toDomainModel
 import br.dev.singular.overview.data.network.source.ISuggestionRemoteDataSource
+import br.dev.singular.overview.data.util.mappers.toDomainModel
 import br.dev.singular.overview.data.util.suggestionModels
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
